+++ ase/test/siesta/test_scripts/H2/script.py
@@ -1,30 +1,28 @@
 from ase.units import Ry
 
 from ase.calculators.siesta import Siesta
-from ase.calculators.siesta.parameters import Specie
+from ase.calculators.siesta.parameters import Specie, PAOBasisBlock
 from ase.optimize import QuasiNewton
 from ase import Atoms
 
-h = Atoms(
+atoms = Atoms(
     '3H',
     [(0.0, 0.0, 0.0),
      (0.0, 0.0, 0.5),
      (0.0, 0.0, 1.0)],
     cell=[10, 10, 10])
 
-h.set_tags([1, 2, 3])
-h.set_initial_magnetic_moments([0, 0, 0])
-
+basis_set = PAOBasisBlock(
+"""1
+0  2 S 0.2
+0.0 0.0""")
+atoms.set_tags([0, 1, 0])
 siesta = Siesta(
-    mesh_cutoff=200 * Ry,
-    basis_set='SZ',
-    spin='COLLINEAR',
-    xc='PBE',
-    pseudo_qualifier='gga',
     species=[
-        Specie(symbol='H', tag=2, basis_set='DZP', ghost=True)],
-    fdf_arguments={'DM.Tolerance': 1e-3})
+        Specie(symbol='H', tag=None, basis_set='SZ'),
+        Specie(symbol='H', tag=1, basis_set=basis_set, ghost=True)],
+    )
 
-h.set_calculator(siesta)
-dyn = QuasiNewton(h, trajectory='h.traj')
+atoms.set_calculator(siesta)
+dyn = QuasiNewton(atoms, trajectory='h.traj')
 dyn.run(fmax=0.02)